+++ dev/dev.js
@@ -1,4 +1,0 @@
-import { createApp, h } from 'vue'
-import Dev from './Dev.vue'
-
-createApp(Dev).mount('#app')