+++ CHANGELOG.md
@@ -2,6 +2,45 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.6](https://github.com/howard-tzw/vue3-select/compare/v0.0.5...v0.0.6) (2023-06-05)
+
+
+### Bug Fixes
+
+* set the pointer to the selected option for the computed options ([b710b8e](https://github.com/howard-tzw/vue3-select/commit/b710b8e9ba98ddaab6a75cdc4475236a3c128bea))
+
+
+### Style Updates
+
+* export --vs-height ([bf2d2cb](https://github.com/howard-tzw/vue3-select/commit/bf2d2cbe644ac8db1a2b8b5feac6bcbd4651809c))
+* selected option highlight in menu and add --vs-dropdown-options-gap ([49dcbc3](https://github.com/howard-tzw/vue3-select/commit/49dcbc3f12bcba086f3bc597e730684a2389ae32))
+
+
+### Chores
+
+* remove scss folder ([1943696](https://github.com/howard-tzw/vue3-select/commit/194369614c48614c92171c5955b351a1fb2a3fc1))
+
+### [0.0.5](https://github.com/howard-tzw/vue3-select/compare/v0.0.4...v0.0.5) (2023-06-02)
+
+
+### Style Updates
+
+* export  --vs-dropdown-border-radius ([48eb5c5](https://github.com/howard-tzw/vue3-select/commit/48eb5c579c3cf8225d40b5985f476af3fee22fb9))
+
+### [0.0.4](https://github.com/howard-tzw/vue3-select/compare/v0.0.3...v0.0.4) (2023-06-02)
+
+
+### Features
+
+* add touchstart event for option hover style on mobile ([b41c719](https://github.com/howard-tzw/vue3-select/commit/b41c719846164773031fa28dd8422b116fb8ec13))
+* update with new css style ([45046b6](https://github.com/howard-tzw/vue3-select/commit/45046b6745101962fad950fbbc858269adc9d81e))
+
+
+### Chores
+
+* remove github issue template ([3e6f55f](https://github.com/howard-tzw/vue3-select/commit/3e6f55f42817f7b5d2e4c96e2ff592f2446cff48))
+* update dev environment ([2090a30](https://github.com/howard-tzw/vue3-select/commit/2090a30a24e6f9615fd9e5bfa70b7b9ace460c02))
+
 ### [0.0.3](https://github.com/howard-tzw/vue3-select/compare/v0.0.2...v0.0.3) (2023-05-09)